move persona to react-internal
