+++ packages/react-internal/src/components/Persona/Persona.base.tsx
@@ -1,7 +1,12 @@
 import * as React from 'react';
-import { warnDeprecations, classNamesFunction, divProperties, getNativeProps, IRenderFunction } from '../../Utilities';
+import {
+  classNamesFunction,
+  divProperties,
+  getNativeProps,
+  IRenderFunction,
+  getPropsWithDefaults,
+} from '../../Utilities';
 import { TooltipHost, TooltipOverflowMode } from '../../Tooltip';
-import { DirectionalHint } from '../../common/DirectionalHint';
 import { PersonaCoin } from './PersonaCoin/PersonaCoin';
 import {
   IPersonaProps,
@@ -11,44 +16,109 @@ import {
   PersonaSize,
   IPersonaCoinProps,
 } from './Persona.types';
+import { useWarnings, useMergedRefs } from '@uifabric/react-hooks';
+import { DirectionalHint } from '../../common/DirectionalHint';
 
 const getClassNames = classNamesFunction<IPersonaStyleProps, IPersonaStyles>();
+
+const DEFAULT_PROPS = {
+  size: PersonaSize.size48,
+  presence: PersonaPresenceEnum.none,
+  imageAlt: '',
+};
+
+function useDebugWarnings(props: IPersonaProps) {
+  if (process.env.NODE_ENV !== 'production') {
+    // eslint-disable-next-line react-hooks/rules-of-hooks -- build-time conditional
+    useWarnings({
+      name: 'Persona',
+      props,
+      deprecations: { primaryText: 'text' },
+    });
+  }
+}
 
 /**
  * Persona with no default styles.
  * [Use the `styles` API to add your own styles.](https://github.com/microsoft/fluentui/wiki/Styling)
  */
-export class PersonaBase extends React.Component<IPersonaProps, {}> {
-  public static defaultProps: IPersonaProps = {
-    size: PersonaSize.size48,
-    presence: PersonaPresenceEnum.none,
-    imageAlt: '',
-  };
+export const PersonaBase: React.FunctionComponent<IPersonaProps> = React.forwardRef<HTMLDivElement, IPersonaProps>(
+  (propsWithoutDefaults, forwardedRef) => {
+    const props = getPropsWithDefaults(DEFAULT_PROPS, propsWithoutDefaults);
 
-  constructor(props: IPersonaProps) {
-    super(props);
+    useDebugWarnings(props);
+    const rootRef = React.useRef<HTMLDivElement>(null);
+    const mergedRootRef = useMergedRefs(forwardedRef, rootRef);
 
-    if (process.env.NODE_ENV !== 'production') {
-      warnDeprecations('Persona', props, { primaryText: 'text' });
-    }
-  }
+    /**
+     * Deprecation helper for getting text.
+     */
+    const getText = (): string => {
+      // eslint-disable-next-line deprecation/deprecation
+      return props.text || props.primaryText || '';
+    };
 
-  public render(): JSX.Element {
-    // wrapping default render behavior based on various this.props properties
-    const _onRenderPrimaryText = this._onRenderText(this._getText());
-    const _onRenderSecondaryText = this._onRenderText(this.props.secondaryText);
-    const _onRenderTertiaryText = this._onRenderText(this.props.tertiaryText);
-    const _onRenderOptionalText = this._onRenderText(this.props.optionalText);
+    /**
+     * Renders various types of Text (primaryText, secondaryText, etc)
+     * based on the classNames passed
+     * @param elementClassNames - element className
+     * @param renderFunction - render function
+     * @param defaultRenderFunction - default render function
+     */
+    const renderElement = (
+      elementClassNames: string,
+      renderFunction: IRenderFunction<IPersonaProps> | undefined,
+      defaultRenderFunction: IRenderFunction<IPersonaProps> | undefined,
+    ): JSX.Element => {
+      return (
+        <div dir="auto" className={elementClassNames}>
+          {renderFunction && renderFunction(props, defaultRenderFunction)}
+        </div>
+      );
+    };
+
+    /**
+     * using closure to wrap the default render behavior
+     * to make it independent of the type of text passed
+     * @param text - text to render
+     */
+    const onRenderText = (text: string | undefined): IRenderFunction<IPersonaProps> | undefined => {
+      // return default render behavior for valid text or undefined
+      return text
+        ? (): JSX.Element => {
+            // default onRender behavior
+            return (
+              <TooltipHost
+                content={text}
+                overflowMode={TooltipOverflowMode.Parent}
+                directionalHint={DirectionalHint.topLeftEdge}
+              >
+                {text}
+              </TooltipHost>
+            );
+          }
+        : undefined;
+    };
+
+    const onInternalRenderPersonaCoin = (providedCoinProps: IPersonaCoinProps): JSX.Element | null => {
+      return <PersonaCoin {...providedCoinProps} />;
+    };
+
+    // wrapping default render behavior based on various props properties
+    const onInternalRenderPrimaryText = onRenderText(getText());
+    const onInternalRenderSecondaryText = onRenderText(props.secondaryText);
+    const onInternalRenderTertiaryText = onRenderText(props.tertiaryText);
+    const onInternalRenderOptionalText = onRenderText(props.optionalText);
 
     const {
       hidePersonaDetails,
-      onRenderOptionalText = _onRenderOptionalText,
-      onRenderPrimaryText = _onRenderPrimaryText,
-      onRenderSecondaryText = _onRenderSecondaryText,
-      onRenderTertiaryText = _onRenderTertiaryText,
-      onRenderPersonaCoin = this._onRenderPersonaCoin,
-    } = this.props;
-    const size = this.props.size as PersonaSize;
+      onRenderOptionalText = onInternalRenderOptionalText,
+      onRenderPrimaryText = onInternalRenderPrimaryText,
+      onRenderSecondaryText = onInternalRenderSecondaryText,
+      onRenderTertiaryText = onInternalRenderTertiaryText,
+      onRenderPersonaCoin = onInternalRenderPersonaCoin,
+    } = props;
+    const size = props.size as PersonaSize;
 
     // These properties are to be explicitly passed into PersonaCoin because they are the only props directly used
     const {
@@ -75,7 +145,7 @@ export class PersonaBase extends React.Component<IPersonaProps, {}> {
       showInitialsUntilImageLoads,
       showSecondaryText,
       theme,
-    } = this.props;
+    } = props;
 
     const personaCoinProps: IPersonaCoinProps = {
       allowPhoneInitials,
@@ -94,7 +164,7 @@ export class PersonaBase extends React.Component<IPersonaProps, {}> {
       presenceTitle,
       showInitialsUntilImageLoads,
       size,
-      text: this._getText(),
+      text: getText(),
       isOutOfOffice,
       presenceColors,
       ...coinProps,
@@ -108,25 +178,27 @@ export class PersonaBase extends React.Component<IPersonaProps, {}> {
       size,
     });
 
-    const divProps = getNativeProps<React.HTMLAttributes<HTMLDivElement>>(this.props, divProperties);
+    const divProps = getNativeProps<React.HTMLAttributes<HTMLDivElement>>(props, divProperties);
     const personaDetails = (
       <div className={classNames.details}>
-        {this._renderElement(classNames.primaryText, onRenderPrimaryText, _onRenderPrimaryText)}
-        {this._renderElement(classNames.secondaryText, onRenderSecondaryText, _onRenderSecondaryText)}
-        {this._renderElement(classNames.tertiaryText, onRenderTertiaryText, _onRenderTertiaryText)}
-        {this._renderElement(classNames.optionalText, onRenderOptionalText, _onRenderOptionalText)}
-        {this.props.children}
+        {renderElement(classNames.primaryText, onRenderPrimaryText, onInternalRenderPrimaryText)}
+        {renderElement(classNames.secondaryText, onRenderSecondaryText, onInternalRenderSecondaryText)}
+        {renderElement(classNames.tertiaryText, onRenderTertiaryText, onInternalRenderTertiaryText)}
+        {renderElement(classNames.optionalText, onRenderOptionalText, onInternalRenderOptionalText)}
+        {props.children}
       </div>
     );
 
     return (
       <div
         {...divProps}
+        ref={mergedRootRef}
         className={classNames.root}
         style={coinSize ? { height: coinSize, minWidth: coinSize } : undefined}
       >
-        {onRenderPersonaCoin(personaCoinProps, this._onRenderPersonaCoin)}
+        {onRenderPersonaCoin(personaCoinProps, onRenderPersonaCoin)}
         {/* eslint-disable deprecation/deprecation */
+
         (!hidePersonaDetails ||
           size === PersonaSize.size8 ||
           size === PersonaSize.size10 ||
@@ -136,59 +208,6 @@ export class PersonaBase extends React.Component<IPersonaProps, {}> {
         }
       </div>
     );
-  }
-
-  /**
-   * Renders various types of Text (primaryText, secondaryText, etc)
-   * based on the classNames passed
-   * @param classNames - element className
-   * @param renderFunction - render function
-   * @param defaultRenderFunction - default render function
-   */
-  private _renderElement(
-    classNames: string,
-    renderFunction: IRenderFunction<IPersonaProps> | undefined,
-    defaultRenderFunction: IRenderFunction<IPersonaProps> | undefined,
-  ): JSX.Element {
-    return (
-      <div dir="auto" className={classNames}>
-        {renderFunction && renderFunction(this.props, defaultRenderFunction)}
-      </div>
-    );
-  }
-
-  /**
-   * Deprecation helper for getting text.
-   */
-  private _getText(): string {
-    // eslint-disable-next-line deprecation/deprecation
-    return this.props.text || this.props.primaryText || '';
-  }
-
-  /**
-   * using closure to wrap the default render behavior
-   * to make it independent of the type of text passed
-   * @param text - text to render
-   */
-  private _onRenderText(text: string | undefined): IRenderFunction<IPersonaProps> | undefined {
-    // return default render behaviour for valid text or undefined
-    return text
-      ? (): JSX.Element => {
-          // default onRender behaviour
-          return (
-            <TooltipHost
-              content={text}
-              overflowMode={TooltipOverflowMode.Parent}
-              directionalHint={DirectionalHint.topLeftEdge}
-            >
-              {text}
-            </TooltipHost>
-          );
-        }
-      : undefined;
-  }
-
-  private _onRenderPersonaCoin = (props: IPersonaCoinProps): JSX.Element | null => {
-    return <PersonaCoin {...props} />;
-  };
-}
+  },
+);
+PersonaBase.displayName = 'PersonaBase';
