+++ packages/react-internal/src/components/Persona/Persona.test.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import * as renderer from 'react-test-renderer';
+import { create } from '@uifabric/utilities/lib/test';
 import { TestImages } from '@uifabric/example-data';
 import { Icon } from '../../Icon';
 import { setRTL, IRenderFunction } from '../../Utilities';
@@ -7,6 +7,7 @@ import { Persona } from './Persona';
 import { mount, ReactWrapper } from 'enzyme';
 import { getIcon } from '../../Styling';
 import { IPersonaSharedProps, IPersonaProps, IPersonaCoinProps, PersonaPresence, PersonaSize } from './index';
+import { isConformant } from '../../common/isConformant';
 
 const testImage1x1 =
   'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAAC0lEQVQImWP4DwQACfsD/eNV8pwAAAAASUVORK5CYII=';
@@ -55,25 +56,25 @@ describe('Persona', () => {
   });
 
   it('renders Persona correctly with no props', () => {
-    const component = renderer.create(<Persona />);
+    const component = create(<Persona />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('renders Persona correctly with initials', () => {
-    const component = renderer.create(<Persona text="Kat Larrson" />);
+    const component = create(<Persona text="Kat Larrson" />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('renders Persona correctly with image', () => {
-    const component = renderer.create(<Persona text="Kat Larrson" imageUrl={testImage1x1} />);
+    const component = create(<Persona text="Kat Larrson" imageUrl={testImage1x1} />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('renders Persona correctly with UnknownPersona coin', () => {
-    const component = renderer.create(<Persona text="Kat Larrson" showUnknownPersonaCoin={true} />);
+    const component = create(<Persona text="Kat Larrson" showUnknownPersonaCoin={true} />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
@@ -81,7 +82,7 @@ describe('Persona', () => {
   it('renders Persona which calls onRenderCoin callback without imageUrl', () => {
     // removing imageUrl prop from example
     const { imageUrl, ...exampleWithoutImage } = examplePersona;
-    const component = renderer.create(
+    const component = create(
       <Persona {...exampleWithoutImage} onRenderCoin={wrapPersona(exampleWithoutImage, true)} />,
     );
     const tree = component.toJSON();
@@ -89,15 +90,13 @@ describe('Persona', () => {
   });
 
   it('renders Persona which calls onRenderPersonaCoin callback with custom render', () => {
-    const component = renderer.create(
-      <Persona {...examplePersona} onRenderPersonaCoin={customOnRenderPersonaFunction} />,
-    );
+    const component = create(<Persona {...examplePersona} onRenderPersonaCoin={customOnRenderPersonaFunction} />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('renders correctly with onRender callback', () => {
-    const component = renderer.create(
+    const component = create(
       <Persona
         {...examplePersona}
         onRenderPrimaryText={wrapPersona(examplePersona)}
@@ -110,13 +109,18 @@ describe('Persona', () => {
   });
 
   it('renders Persona children correctly', () => {
-    const component = renderer.create(
+    const component = create(
       <Persona text="Kat Larrson">
         <span>Persona Children</span>
       </Persona>,
     );
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
+  });
+
+  isConformant({
+    Component: Persona,
+    displayName: 'Persona',
   });
 
   describe('initials and colors', () => {
@@ -187,14 +191,14 @@ describe('Persona', () => {
   describe('image', () => {
     it('renders empty alt text by default', () => {
       const wrapper = mount(<Persona text="Kat Larrson" imageUrl={testImage1x1} />);
-      const image: ReactWrapper<React.ImgHTMLAttributes<any>, any> = wrapper.find('ImageBase');
+      const image: ReactWrapper<React.ImgHTMLAttributes<unknown>, unknown> = wrapper.find('ImageBase');
 
       expect(image.props().alt).toEqual('');
     });
 
     it('renders its given alt text', () => {
       const wrapper = mount(<Persona text="Kat Larrson" imageUrl={testImage1x1} imageAlt="ALT TEXT" />);
-      const image: ReactWrapper<React.ImgHTMLAttributes<any>, any> = wrapper.find('ImageBase');
+      const image: ReactWrapper<React.ImgHTMLAttributes<unknown>, unknown> = wrapper.find('ImageBase');
 
       expect(image.props().alt).toEqual('ALT TEXT');
     });
