+++ packages/react-internal/src/components/Persona/PersonaCoin/PersonaCoin.base.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react';
 import {
-  warnDeprecations,
   classNamesFunction,
   divProperties,
   getInitials,
   getNativeProps,
   getRTL,
+  getPropsWithDefaults,
 } from '../../../Utilities';
 import { mergeStyles } from '../../../Styling';
 import { PersonaPresence } from '../PersonaPresence/index';
@@ -21,6 +21,7 @@ import {
 } from '../Persona.types';
 import { getPersonaInitialsColor } from '../PersonaInitialsColor';
 import { sizeToPixels } from '../PersonaConsts';
+import { useWarnings } from '@uifabric/react-hooks';
 
 const getClassNames = classNamesFunction<IPersonaCoinStyleProps, IPersonaCoinStyles>({
   // There can be many PersonaCoin rendered with different sizes.
@@ -28,199 +29,191 @@ const getClassNames = classNamesFunction<IPersonaCoinStyleProps, IPersonaCoinSty
   cacheSize: 100,
 });
 
-export interface IPersonaState {
-  isImageLoaded?: boolean;
-  isImageError?: boolean;
+const DEFAULT_PROPS = {
+  size: PersonaSize.size48,
+  presence: PersonaPresenceEnum.none,
+  imageAlt: '',
+};
+
+function useDebugWarnings(props: IPersonaCoinProps) {
+  if (process.env.NODE_ENV !== 'production') {
+    // eslint-disable-next-line react-hooks/rules-of-hooks -- build-time conditional
+    useWarnings({
+      name: 'PersonaCoin',
+      props,
+      deprecations: { primaryText: 'text' },
+    });
+  }
+}
+
+function useImageLoadState({ onPhotoLoadingStateChange, imageUrl }: IPersonaCoinProps) {
+  const [imageLoadState, setImageLoadstate] = React.useState<ImageLoadState>(ImageLoadState.notLoaded);
+
+  React.useEffect(() => {
+    setImageLoadstate(ImageLoadState.notLoaded);
+  }, [imageUrl]);
+
+  const onLoadingStateChange = (loadState: ImageLoadState) => {
+    setImageLoadstate(loadState);
+
+    onPhotoLoadingStateChange?.(loadState);
+  };
+
+  return [imageLoadState, onLoadingStateChange] as const;
 }
 
 /**
  * PersonaCoin with no default styles.
  * [Use the `getStyles` API to add your own styles.](https://github.com/microsoft/fluentui/wiki/Styling)
  */
-export class PersonaCoinBase extends React.Component<IPersonaCoinProps, IPersonaState> {
-  public static defaultProps: IPersonaCoinProps = {
-    size: PersonaSize.size48,
-    presence: PersonaPresenceEnum.none,
-    imageAlt: '',
-  };
+export const PersonaCoinBase: React.FunctionComponent<IPersonaCoinProps> = React.forwardRef<
+  HTMLDivElement,
+  IPersonaCoinProps
+>((propsWithoutDefaults, forwardedRef) => {
+  const props = getPropsWithDefaults(DEFAULT_PROPS, propsWithoutDefaults);
 
-  constructor(props: IPersonaCoinProps) {
-    super(props);
+  useDebugWarnings(props);
 
-    if (process.env.NODE_ENV !== 'production') {
-      warnDeprecations('PersonaCoin', props, { primaryText: 'text' });
-    }
+  const [imageLoadState, onLoadingStateChange] = useImageLoadState(props);
 
-    this.state = {
-      isImageLoaded: false,
-      isImageError: false,
-    };
-  }
+  const renderCoin = getCoinRenderer(onLoadingStateChange);
 
-  public UNSAFE_componentWillReceiveProps(nextProps: IPersonaCoinProps): void {
-    if (nextProps.imageUrl !== this.props.imageUrl) {
-      this.setState({
-        isImageLoaded: false,
-        isImageError: false,
-      });
-    }
-  }
-
-  public render(): JSX.Element | null {
-    const {
-      className,
-      coinProps,
-      showUnknownPersonaCoin,
-      coinSize,
-      styles,
-      imageUrl,
-      isOutOfOffice,
-      /* eslint-disable deprecation/deprecation */
-      onRenderCoin = this._onRenderCoin,
-      onRenderPersonaCoin = onRenderCoin,
-      /* eslint-enable deprecation/deprecation */
-      onRenderInitials = this._onRenderInitials,
-      presence,
-      presenceTitle,
-      presenceColors,
-      showInitialsUntilImageLoads,
-      theme,
-    } = this.props;
-
-    const size = this.props.size as PersonaSize;
-    const divProps = getNativeProps<React.HTMLAttributes<HTMLDivElement>>(this.props, divProperties);
-    const divCoinProps = getNativeProps<React.HTMLAttributes<HTMLDivElement>>(coinProps || {}, divProperties);
-    const coinSizeStyle = coinSize ? { width: coinSize, height: coinSize } : undefined;
-    const hideImage = showUnknownPersonaCoin;
-
-    const personaPresenceProps: IPersonaPresenceProps = {
-      coinSize,
-      isOutOfOffice,
-      presence,
-      presenceTitle,
-      presenceColors,
-      size,
-      theme,
-    };
-
-    // Use getStyles from props, or fall back to getStyles from styles file.
-    const classNames = getClassNames(styles, {
-      theme: theme!,
-      className: coinProps && coinProps.className ? coinProps.className : className,
-      size,
-      coinSize,
-      showUnknownPersonaCoin,
-    });
-
-    const shouldRenderInitials = Boolean(
-      !this.state.isImageLoaded &&
-        ((showInitialsUntilImageLoads && imageUrl) || !imageUrl || this.state.isImageError || hideImage),
-    );
-
-    return (
-      <div role="presentation" {...divProps} className={classNames.coin}>
-        {// Render PersonaCoin if size is not size8. size10 and tiny need to removed after a deprecation cleanup.
-        // eslint-disable-next-line deprecation/deprecation
-        size !== PersonaSize.size8 && size !== PersonaSize.size10 && size !== PersonaSize.tiny ? (
-          <div role="presentation" {...divCoinProps} className={classNames.imageArea} style={coinSizeStyle}>
-            {shouldRenderInitials && (
-              <div
-                className={mergeStyles(
-                  classNames.initials,
-                  !showUnknownPersonaCoin && { backgroundColor: getPersonaInitialsColor(this.props) },
-                )}
-                style={coinSizeStyle}
-                aria-hidden="true"
-              >
-                {onRenderInitials(this.props, this._onRenderInitials)}
-              </div>
-            )}
-            {!hideImage && onRenderPersonaCoin(this.props, this._onRenderCoin)}
-            <PersonaPresence {...personaPresenceProps} />
-          </div>
-        ) : // Otherwise, render just PersonaPresence.
-        this.props.presence ? (
-          <PersonaPresence {...personaPresenceProps} />
-        ) : (
-          // Just render Contact Icon if there isn't a Presence prop.
-          <Icon iconName="Contact" className={classNames.size10WithoutPresenceIcon} />
-        )}
-        {this.props.children}
-      </div>
-    );
-  }
-
-  private _onRenderCoin = (props: IPersonaCoinProps): JSX.Element | null => {
-    const {
-      coinSize,
-      styles,
-      imageUrl,
-      imageAlt,
-      imageShouldFadeIn,
-      imageShouldStartVisible,
-      theme,
-      showUnknownPersonaCoin,
-    } = this.props;
-
-    // Render the Image component only if an image URL is provided
-    if (!imageUrl) {
-      return null;
-    }
-
-    const size = this.props.size as PersonaSize;
-
-    const classNames = getClassNames(styles, {
-      theme: theme!,
-      size,
-      showUnknownPersonaCoin,
-    });
-
-    const dimension = coinSize || sizeToPixels[size];
-
-    return (
-      <Image
-        className={classNames.image}
-        imageFit={ImageFit.cover}
-        src={imageUrl}
-        width={dimension}
-        height={dimension}
-        alt={imageAlt}
-        shouldFadeIn={imageShouldFadeIn}
-        shouldStartVisible={imageShouldStartVisible}
-        onLoadingStateChange={this._onPhotoLoadingStateChange}
-      />
-    );
-  };
-
-  /**
-   * Deprecation helper for getting text.
-   */
-  private _getText(): string {
+  const {
+    className,
+    coinProps,
+    showUnknownPersonaCoin,
+    coinSize,
+    styles,
+    imageUrl,
+    isOutOfOffice,
     // eslint-disable-next-line deprecation/deprecation
-    return this.props.text || this.props.primaryText || '';
+    onRenderCoin = renderCoin,
+    // eslint-disable-next-line deprecation/deprecation
+    onRenderPersonaCoin = onRenderCoin,
+    onRenderInitials = renderPersonaCoinInitials,
+    presence,
+    presenceTitle,
+    presenceColors,
+    showInitialsUntilImageLoads,
+    theme,
+    size,
+  } = props;
+
+  const divProps = getNativeProps<React.HTMLAttributes<HTMLDivElement>>(props, divProperties);
+  const divCoinProps = getNativeProps<React.HTMLAttributes<HTMLDivElement>>(coinProps || {}, divProperties);
+  const coinSizeStyle = coinSize ? { width: coinSize, height: coinSize } : undefined;
+  const hideImage = showUnknownPersonaCoin;
+
+  const personaPresenceProps: IPersonaPresenceProps = {
+    coinSize,
+    isOutOfOffice,
+    presence,
+    presenceTitle,
+    presenceColors,
+    size,
+    theme,
+  };
+
+  // Use getStyles from props, or fall back to getStyles from styles file.
+  const classNames = getClassNames(styles, {
+    theme: theme!,
+    className: coinProps && coinProps.className ? coinProps.className : className,
+    size,
+    coinSize,
+    showUnknownPersonaCoin,
+  });
+
+  const shouldRenderInitials = Boolean(
+    imageLoadState !== ImageLoadState.loaded &&
+      ((showInitialsUntilImageLoads && imageUrl) || !imageUrl || imageLoadState === ImageLoadState.error || hideImage),
+  );
+
+  return (
+    <div role="presentation" {...divProps} className={classNames.coin} ref={forwardedRef}>
+      {// Render PersonaCoin if size is not size8. size10 and tiny need to removed after a deprecation cleanup.
+      // eslint-disable-next-line deprecation/deprecation
+      size !== PersonaSize.size8 && size !== PersonaSize.size10 && size !== PersonaSize.tiny ? (
+        <div role="presentation" {...divCoinProps} className={classNames.imageArea} style={coinSizeStyle}>
+          {shouldRenderInitials && (
+            <div
+              className={mergeStyles(
+                classNames.initials,
+                !showUnknownPersonaCoin && { backgroundColor: getPersonaInitialsColor(props) },
+              )}
+              style={coinSizeStyle}
+              aria-hidden="true"
+            >
+              {onRenderInitials(props, renderPersonaCoinInitials)}
+            </div>
+          )}
+          {!hideImage && onRenderPersonaCoin(props, renderCoin)}
+          <PersonaPresence {...personaPresenceProps} />
+        </div>
+      ) : // Otherwise, render just PersonaPresence.
+      props.presence ? (
+        <PersonaPresence {...personaPresenceProps} />
+      ) : (
+        // Just render Contact Icon if there isn't a Presence prop.
+        <Icon iconName="Contact" className={classNames.size10WithoutPresenceIcon} />
+      )}
+      {props.children}
+    </div>
+  );
+});
+PersonaCoinBase.displayName = 'PersonaCoinBase';
+
+const getCoinRenderer = (onLoadingStateChange: (loadState: ImageLoadState) => void) => ({
+  coinSize,
+  styles,
+  imageUrl,
+  imageAlt,
+  imageShouldFadeIn,
+  imageShouldStartVisible,
+  theme,
+  showUnknownPersonaCoin,
+  size = DEFAULT_PROPS.size,
+}: IPersonaCoinProps): JSX.Element | null => {
+  // Render the Image component only if an image URL is provided
+  if (!imageUrl) {
+    return null;
+  }
+  const classNames = getClassNames(styles, {
+    theme: theme!,
+    size,
+    showUnknownPersonaCoin,
+  });
+  const dimension = coinSize || sizeToPixels[size];
+  return (
+    <Image
+      className={classNames.image}
+      imageFit={ImageFit.cover}
+      src={imageUrl}
+      width={dimension}
+      height={dimension}
+      alt={imageAlt}
+      shouldFadeIn={imageShouldFadeIn}
+      shouldStartVisible={imageShouldStartVisible}
+      onLoadingStateChange={onLoadingStateChange}
+    />
+  );
+};
+
+const renderPersonaCoinInitials = ({
+  imageInitials,
+  allowPhoneInitials,
+  showUnknownPersonaCoin,
+  text,
+  // eslint-disable-next-line deprecation/deprecation
+  primaryText,
+  theme,
+}: IPersonaCoinProps): JSX.Element => {
+  if (showUnknownPersonaCoin) {
+    return <Icon iconName="Help" />;
   }
 
-  private _onRenderInitials = (props: IPersonaCoinProps): JSX.Element => {
-    let { imageInitials } = props;
-    const { allowPhoneInitials, showUnknownPersonaCoin } = props;
+  const isRTL = getRTL(theme);
 
-    if (showUnknownPersonaCoin) {
-      return <Icon iconName="Help" />;
-    }
+  imageInitials = imageInitials || getInitials(text || primaryText || '', isRTL, allowPhoneInitials);
 
-    const isRTL = getRTL(this.props.theme);
-
-    imageInitials = imageInitials || getInitials(this._getText(), isRTL, allowPhoneInitials);
-
-    return imageInitials !== '' ? <span>{imageInitials}</span> : <Icon iconName="Contact" />;
-  };
-
-  private _onPhotoLoadingStateChange = (loadState: ImageLoadState) => {
-    this.setState({
-      isImageLoaded: loadState === ImageLoadState.loaded,
-      isImageError: loadState === ImageLoadState.error,
-    });
-
-    this.props.onPhotoLoadingStateChange && this.props.onPhotoLoadingStateChange(loadState);
-  };
-}
+  return imageInitials !== '' ? <span>{imageInitials}</span> : <Icon iconName="Contact" />;
+};
