+++ packages/react-internal/src/components/Persona/PersonaCoin/PersonaCoin.deprecated.test.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import * as renderer from 'react-test-renderer';
+import { create } from '@uifabric/utilities/lib/test';
 import { setWarningCallback, setRTL } from '../../../Utilities';
 import { PersonaCoin } from './PersonaCoin';
 
@@ -23,25 +23,25 @@ describe('PersonaCoin', () => {
   });
 
   it('renders correctly', () => {
-    const component = renderer.create(<PersonaCoin />);
+    const component = create(<PersonaCoin />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('renders correctly with text', () => {
-    const component = renderer.create(<PersonaCoin primaryText="Kat Larrson" />);
+    const component = create(<PersonaCoin primaryText="Kat Larrson" />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('renders correctly with provided initials', () => {
-    const component = renderer.create(<PersonaCoin imageInitials="JG" />);
+    const component = create(<PersonaCoin imageInitials="JG" />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('renders correctly with image', () => {
-    const component = renderer.create(<PersonaCoin primaryText="Kat Larrson" imageUrl={testImage1x1} />);
+    const component = create(<PersonaCoin primaryText="Kat Larrson" imageUrl={testImage1x1} />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
