+++ packages/react-internal/src/components/Persona/PersonaPresence/PersonaPresence.test.tsx
@@ -1,8 +1,9 @@
 import * as React from 'react';
-import * as renderer from 'react-test-renderer';
+import { create } from '@uifabric/utilities/lib/test';
 import { setRTL } from '../../../Utilities';
 import * as PersonaTypes from '../Persona.types';
 import { PersonaPresence } from './PersonaPresence';
+import { isConformant } from '../../../common/isConformant';
 
 describe('PersonaPresence', () => {
   beforeEach(() => {
@@ -10,91 +11,93 @@ describe('PersonaPresence', () => {
   });
 
   it('renders available', () => {
-    const component = renderer.create(<PersonaPresence presence={PersonaTypes.PersonaPresence.online} />);
+    const component = create(<PersonaPresence presence={PersonaTypes.PersonaPresence.online} />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('renders available + out of office', () => {
-    const component = renderer.create(<PersonaPresence presence={PersonaTypes.PersonaPresence.online} isOutOfOffice />);
+    const component = create(<PersonaPresence presence={PersonaTypes.PersonaPresence.online} isOutOfOffice />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('renders away', () => {
-    const component = renderer.create(<PersonaPresence presence={PersonaTypes.PersonaPresence.away} />);
+    const component = create(<PersonaPresence presence={PersonaTypes.PersonaPresence.away} />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('renders away + out of office', () => {
-    const component = renderer.create(<PersonaPresence presence={PersonaTypes.PersonaPresence.away} isOutOfOffice />);
+    const component = create(<PersonaPresence presence={PersonaTypes.PersonaPresence.away} isOutOfOffice />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('renders busy', () => {
-    const component = renderer.create(<PersonaPresence presence={PersonaTypes.PersonaPresence.busy} />);
+    const component = create(<PersonaPresence presence={PersonaTypes.PersonaPresence.busy} />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('renders busy + out of office', () => {
-    const component = renderer.create(<PersonaPresence presence={PersonaTypes.PersonaPresence.busy} isOutOfOffice />);
+    const component = create(<PersonaPresence presence={PersonaTypes.PersonaPresence.busy} isOutOfOffice />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('renders do not disturb', () => {
-    const component = renderer.create(<PersonaPresence presence={PersonaTypes.PersonaPresence.dnd} />);
+    const component = create(<PersonaPresence presence={PersonaTypes.PersonaPresence.dnd} />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('renders do not disturb + out of office', () => {
-    const component = renderer.create(<PersonaPresence presence={PersonaTypes.PersonaPresence.dnd} isOutOfOffice />);
+    const component = create(<PersonaPresence presence={PersonaTypes.PersonaPresence.dnd} isOutOfOffice />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('renders blocked', () => {
-    const component = renderer.create(<PersonaPresence presence={PersonaTypes.PersonaPresence.blocked} />);
+    const component = create(<PersonaPresence presence={PersonaTypes.PersonaPresence.blocked} />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('renders blocked + out of office', () => {
     // Blocked + out of office does not exist and is the same as regular blocked
-    const component = renderer.create(
-      <PersonaPresence presence={PersonaTypes.PersonaPresence.blocked} isOutOfOffice />,
-    );
+    const component = create(<PersonaPresence presence={PersonaTypes.PersonaPresence.blocked} isOutOfOffice />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('renders offline', () => {
-    const component = renderer.create(<PersonaPresence presence={PersonaTypes.PersonaPresence.offline} />);
+    const component = create(<PersonaPresence presence={PersonaTypes.PersonaPresence.offline} />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('renders offline + out of office', () => {
-    const component = renderer.create(
-      <PersonaPresence presence={PersonaTypes.PersonaPresence.offline} isOutOfOffice />,
-    );
+    const component = create(<PersonaPresence presence={PersonaTypes.PersonaPresence.offline} isOutOfOffice />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('renders none', () => {
-    const component = renderer.create(<PersonaPresence presence={PersonaTypes.PersonaPresence.none} />);
+    const component = create(<PersonaPresence presence={PersonaTypes.PersonaPresence.none} />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('renders none + out of office', () => {
-    const component = renderer.create(<PersonaPresence presence={PersonaTypes.PersonaPresence.none} isOutOfOffice />);
+    const component = create(<PersonaPresence presence={PersonaTypes.PersonaPresence.none} isOutOfOffice />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
+  });
+
+  isConformant({
+    Component: PersonaPresence,
+    displayName: 'PersonaPresence',
+    disabledTests: ['exported-top-level', 'has-top-level-file'],
   });
 });
