--- packages/react-next/src/components/Persona/PersonaCoin/index.ts
@@ -1,2 +1,0 @@
-export * from './PersonaCoin';
-export * from './PersonaCoin.base';
