--- packages/react-next/src/components/Persona/PersonaPresence/index.ts
@@ -1,2 +1,0 @@
-export * from './PersonaPresence';
-export * from './PersonaPresence.base';
